Added popAll and Changed getOutput
I added popAll to Stack.c and .h, but it is broken and needs to be
fixed. I changed getOutput to accept an additional parameter that is a
double* called output that will be the answer of the expression. Now
getOutput will return 0 if there was a problem with the formula and 1 if
there was not. This was getOutput will check if the reverse polish
notation formula is proper and obtains the answer.